Phase 2: fix pci.h bitfield base-type bug
struct pci_regs declared the BIST byte as 'u8 bist :8' among three
'u32 ...:8' siblings in the same 32-bit register. GCC packs mixed-
base-type bitfields tight (168 bytes), but ctypes treats the u8->u32
type change as a storage-unit break (169) -- so the regenerated
tglib.py struct disagreed with the hardware layout and rflip.py's
size-assertion failed.

Changed bist to u32, matching its siblings. tglib.py is byte-exact
again and the entire core import graph (device.py and everything it
pulls in) now imports cleanly on Python 3.

+++ include/pci.h
@@ -170,7 +170,7 @@ struct pci_regs {
     };
     struct pci_class_code_rev_id class_code_rev_id;
     struct {
-        u8 bist :8;
+        u32 bist :8;
         u32 hdr_type :8;
         u32 lat_timer :8;
         u32 cache_line_sz :8;
